+++ bin/rdme
@@ -1,29 +1,29 @@
 #! /usr/bin/env node
+/* eslint-disable no-console */
 const chalk = require('chalk');
 const core = require('@actions/core');
 
-const nodeVersion = require('parse-node-version')(process.version);
 const updateNotifier = require('update-notifier');
 const pkg = require('../package.json');
 
 const isGHA = require('../src/lib/isGitHub');
+const { isSupportedNodeVersion } = require('../src/lib/nodeVersionUtils');
 
 updateNotifier({ pkg }).notify();
 
 /**
  * We use optional chaining throughout the library, which doesn't work on Node 12, so to curb
  * support questions about why rdme is throwing an "Unexpected token '.'" error we should hard
- * stop if we're on any Node version that doesn't support it.
+ * stop if we're being run with any Node version that we don't explicitly support.
  */
-if (nodeVersion.major <= 12) {
-  const message = `We're sorry, this release of rdme does not support your Node version of ${nodeVersion.major}.`;
+if (!isSupportedNodeVersion(process.version)) {
+  const message = `We're sorry, this release of rdme does not support your Node version of ${process.version}. We support the following versions: ${pkg.engines.node}`;
   console.error(chalk.red(`\n${message}\n`));
   process.exit(1);
 }
 
 require('../src')(process.argv.slice(2))
   .then(msg => {
-    // eslint-disable-next-line no-console
     if (msg) console.log(msg);
     return process.exit(0);
   })
@@ -38,14 +38,14 @@ require('../src')(process.argv.slice(2))
 
     /**
      * If we're in a GitHub Actions environment, log errors with that formatting instead.
-     * @link: https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#setting-an-error-message
-     * @link: https://github.com/actions/toolkit/tree/main/packages/core#annotations
+     *
+     * @see {@link https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#setting-an-error-message}
+     * @see {@link https://github.com/actions/toolkit/tree/main/packages/core#annotations}
      */
     if (isGHA()) {
       return core.setFailed(message);
     }
 
-    // eslint-disable-next-line no-console
     console.error(chalk.red(`\n${message}\n`));
     return process.exit(1);
   });
