+++ package.json
@@ -51,7 +51,6 @@
     "oas-normalize": "^5.2.0",
     "open": "^8.2.1",
     "parse-link-header": "^2.0.0",
-    "parse-node-version": "^2.0.0",
     "read": "^1.0.7",
     "semver": "^7.0.0",
     "tmp-promise": "^3.0.2",
