--- src/lib/getNodeVersion.js
@@ -1,15 +1,0 @@
-const pkg = require('../../package.json');
-
-/**
- * @example 14
- * @returns {String} The maximum major Node.js version specified in the package.json
- */
-function extractNodeVersion() {
-  const { node } = pkg.engines;
-
-  const match = Array.from(node.matchAll(/\d+/g)).pop();
-
-  return match;
-}
-
-module.exports = extractNodeVersion;
